requirements.txt
altgraph==0.17.4
et-xmlfile==1.1.0
macholib==1.16.3
openpyxl==3.1.2
packaging==24.0
pillow==10.2.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
pyserial==3.5
setuptools==69.2.0
webcolors==1.13
